0.19.1
- Fix auto-purging feature and set it to purge records in 1-hour chunks

+++ src/main/java/org/dsa/iot/etsdb/db/Db.java
@@ -70,7 +70,7 @@ public class Db extends Database {
 
     private void setDiskSpaceRemaining(int space) {
         long totalSize = fPath.getTotalSpace();
-        this.diskSpaceRemaining = totalSize * (space / 100);
+        this.diskSpaceRemaining = (long) (totalSize * (space / 100.0));
     }
 
     @Override
